requirements.txt
Flask
gunicorn
jsonify
flask-sqlalchemy
psycopg2-binary
python-dotenv
sqlalchemy